+++ src/Subjects/UrlKeyAwareSubjectInterface.php
@@ -1,0 +1,62 @@
+<?php
+
+/**
+ * TechDivision\Import\Subjects\UrlKeyAwareSubjectInterface
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Open Software License (OSL 3.0)
+ * that is available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/osl-3.0.php
+ *
+ * PHP version 5
+ *
+ * @author    Tim Wagner <t.wagner@techdivision.com>
+ * @copyright 2019 TechDivision GmbH <info@techdivision.com>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/techdivision/import
+ * @link      http://www.techdivision.com
+ */
+
+namespace TechDivision\Import\Subjects;
+
+/**
+ * The interface for all URL key aware subject implementations.
+ *
+ * @author    Tim Wagner <t.wagner@techdivision.com>
+ * @copyright 2019 TechDivision GmbH <info@techdivision.com>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/techdivision/import
+ * @link      http://www.techdivision.com
+ */
+interface UrlKeyAwareSubjectInterface
+{
+
+    /**
+     * Return's the entity type for the configured entity type code.
+     *
+     * @return array The requested entity type
+     * @throws \Exception Is thrown, if the requested entity type is not available
+     */
+    public function getEntityType();
+
+    /**
+     * Return's the store ID of the actual row, or of the default store
+     * if no store view code is set in the CSV file.
+     *
+     * @param string|null $default The default store view code to use, if no store view code is set in the CSV file
+     *
+     * @return integer The ID of the actual store
+     * @throws \Exception Is thrown, if the store with the actual code is not available
+     */
+    public function getRowStoreId($default = null);
+
+    /**
+     * Return's TRUE, if the passed URL key varchar value IS related with the actual PK.
+     *
+     * @param array $varcharAttribute The varchar value to check
+     *
+     * @return boolean TRUE if the URL key is related, else FALSE
+     */
+    public function isUrlKeyOf(array $varcharAttribute);
+}