+++ src/Utils/UrlKeyUtilInterface.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * TechDivision\Import\Configuration\ListenerAwareConfigurationInterface
+ * TechDivision\Import\Utils\UrlKeyUtilInterface
  *
  * NOTICE OF LICENSE
  *
@@ -12,30 +12,34 @@
  * PHP version 5
  *
  * @author    Tim Wagner <t.wagner@techdivision.com>
- * @copyright 2016 TechDivision GmbH <info@techdivision.com>
+ * @copyright 2019 TechDivision GmbH <info@techdivision.com>
  * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  * @link      https://github.com/techdivision/import
  * @link      http://www.techdivision.com
  */
 
-namespace TechDivision\Import\Configuration;
+namespace TechDivision\Import\Utils;
+
+use TechDivision\Import\Subjects\UrlKeyAwareSubjectInterface;
 
 /**
- * The interface for the all configurations that are listener aware.
+ * Interface for utility implementations that provides functionality to make URL keys unique.
  *
  * @author    Tim Wagner <t.wagner@techdivision.com>
- * @copyright 2016 TechDivision GmbH <info@techdivision.com>
+ * @copyright 2019 TechDivision GmbH <info@techdivision.com>
  * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  * @link      https://github.com/techdivision/import
  * @link      http://www.techdivision.com
  */
-interface ListenerAwareConfigurationInterface
+interface UrlKeyUtilInterface
 {
 
     /**
-     * Return's the array with the configured listeners.
+     * Make's the passed URL key unique by adding the next number to the end.
      *
-     * @return array The array with the listeners
+     * @param string $urlKey The URL key to make unique
+     *
+     * @return string The unique URL key
      */
-    public function getListeners();
+    public function makeUnique(UrlKeyAwareSubjectInterface $subject, $urlKey);
 }